[update to 1.25.12] (oss-fuzz-bug 15975). The earlier fix around the same location needed
1.25.12
-------

More credit to OSS-Fuzz. The ID3v2 parser code is not yet as hardened
as the actual MPEG decoder. The paranoid can disable it at build-time.
If you do not need it, this is a good idea, anyway: Code that is not
there, cannot be exploited. Speaking about exploits: The recent crop
of bugs trigger a denial of service (crash) worst-case, some invalid
ID3 data normally. Code injection maybe not totally ruled out (that one
write of a zero byte?), but does not seem easy. Update to be sure that
you are only suceptible to as of yet hidden bugs.

- libmpg123
-- Fix an out-of-bounds read of maximal two bytes for truncated RVA2 frames
   (oss-fuzz-bug 15975). The earlier fix around the same location needed

(NEWS truncated at 15 lines)